feat!: add useGestureViewerState hook and refactor controller
- Add useGestureViewerState hook to access currentIndex and totalCount
- Refactor useGestureViewerController to return control methods only
- Rename GestureViewerControllerState to GestureViewerState
- Update exports and type definitions
