+++ src/index.tsx
@@ -1,11 +1,12 @@
 export { GestureViewer } from './GestureViewer';
 export type {
   GestureViewerController,
-  GestureViewerControllerState,
   GestureViewerEventCallback,
   GestureViewerEventData,
   GestureViewerEventType,
   GestureViewerProps,
+  GestureViewerState,
 } from './types';
 export { useGestureViewerController } from './useGestureViewerController';
 export { useGestureViewerEvent } from './useGestureViewerEvent';
+export { useGestureViewerState } from './useGestureViewerState';
