+++ src/types.ts
@@ -252,13 +252,13 @@ export type GestureViewerController = {
    * ```
    */
   rotate: (angle?: RotationAngle, clockwise?: boolean) => void;
-} & GestureViewerControllerState;
+};
 
 /**
  * State information for the gesture viewer controller.
  * Contains read-only properties that reflect the current state.
  */
-export type GestureViewerControllerState = {
+export type GestureViewerState = {
   /**
    * The current index of the active item in the viewer.
    *
@@ -267,7 +267,7 @@ export type GestureViewerControllerState = {
    *
    * @example
    * ```typescript
-   * console.log(`Currently viewing item ${controller.currentIndex + 1} of ${controller.totalCount}`);
+   * console.log(`Currently viewing item ${currentIndex + 1} of ${totalCount}`);
    * ```
    */
   readonly currentIndex: number;
@@ -280,8 +280,8 @@ export type GestureViewerControllerState = {
    *
    * @example
    * ```typescript
-   * const hasNext = controller.currentIndex < controller.totalCount - 1;
-   * const hasPrevious = controller.currentIndex > 0;
+   * const hasNext = currentIndex < totalCount - 1;
+   * const hasPrevious = currentIndex > 0;
    * ```
    */
   readonly totalCount: number;
